Make it work (almost)
Completely rewritten the extension. The only way I found to make it reliable was to loop through each editors children and map them to placeholders specified in options.

Then, for each child: test if it's present, if it matches the placeholder tagName, if it is not empty, and so on ; then show/hide or create/remove the placeholder element.

Bonus: pressing enter key at the end of current line when the next element is an empty placeholder will focus it instead of inserting a new paragraph.

Works great when having a first heading placeholder, and then a paragraph (like on Medium). Other cases can break sometimes...

Known issue with the `disableDoubleReturn: true` option: the user can't create a new line as long as a empty placeholder is remaining.